refactor(connd): use faux for mocking modem manager
